+++ easypost/easypost_client.py
@@ -5,6 +5,10 @@ from easypost.constant import (
     SUPPORT_EMAIL,
     TIMEOUT,
 )
+from easypost.hooks import (
+    RequestHook,
+    ResponseHook,
+)
 from easypost.services import (
     AddressService,
     BatchService,
@@ -75,6 +79,10 @@ class EasyPostClient:
         self.user = UserService(self)
         self.webhook = WebhookService(self)
 
+        # Hooks
+        self._request_event = RequestHook()
+        self._response_event = ResponseHook()
+
         # use urlfetch as request_lib on google app engine, otherwise use requests
         self._request_lib = None
         try:
@@ -105,3 +113,19 @@ class EasyPostClient:
             else:
                 if major_version < 1:
                     raise ImportError(INVALID_REQUESTS_VERSION_ERROR.format(SUPPORT_EMAIL))
+
+    def subscribe_to_request_hook(self, function):
+        """Subscribe functions to run when a request event occurs."""
+        self._request_event += function
+
+    def unsubscribe_from_request_hook(self, function):
+        """Unsubscribe functions from running when a request even occurs."""
+        self._request_event -= function
+
+    def subscribe_to_response_hook(self, function):
+        """Subscribe functions to run when a response event occurs."""
+        self._response_event += function
+
+    def unsubscribe_from_response_hook(self, function):
+        """Unsubscribe functions from running when a response even occurs."""
+        self._response_event -= function